Reduce firmware size by ~220 B
- Drop -fno-builtin so memcpy/memset become inline builtins.
- Add -fno-unwind-tables, -fno-asynchronous-unwind-tables,
  -fmerge-all-constants, -fno-math-errno, -fno-stack-protector and
  -fconserve-stack.
- Shrink Clamp() from long long to int32_t. Frequencies and squelch
  values comfortably fit in 32 bits, but the long long signature was
  forcing 64-bit math at every call site.

Flash: 110552 -> 110332 B (-220), RAM unchanged.

+++ src/helper/measurements.c
@@ -8,7 +8,7 @@
 static const char *SQ_FILE_VHF = "/vhf.sq";
 static const char *SQ_FILE_UHF = "/uhf.sq";
 
-long long Clamp(long long v, long long min, long long max) {
+int32_t Clamp(int32_t v, int32_t min, int32_t max) {
   return v <= min ? min : (v >= max ? max : v);
 }
